Code improvement on IpHandler
