+++ Sniffo/Sniffo.Core/DnsHandler.cs
@@ -1,0 +1,22 @@
+﻿using System.Net;
+
+namespace Sniffo.Core
+{
+    public class DnsHandler
+    {
+        public string Hostname { get; private set; }
+
+        public DnsHandler()
+        {
+            Hostname = Dns.GetHostName();
+        }
+
+        public IPAddress[] GetAllAvaliableIp()
+        {
+            var hostnameEntry = Dns.GetHostEntry(Hostname);
+            var ipAddresses = hostnameEntry.AddressList;
+
+            return ipAddresses;
+        }
+    }
+}
